% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesort.R
\name{cutbin}
\alias{cutbin}
\title{Cut Into Bins}
\usage{
cutbin(x, breaks, bin_names)
}
\arguments{
\item{x}{column to be cut}

\item{breaks}{define the bins}

\item{bin_names}{names for the result}
}
\value{
bins factor
}
\description{
No boundaries on the endpoints, and handles character \code{x}.
A little different than normal \code{\link[base]{cut}}.
}
