% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesort.R
\name{disksort}
\alias{disksort}
\title{Sort File On Disk}
\usage{
disksort(infile, outfile = NULL, sortcolumn = 1L, nrows = NULL,
  read.table.args = NULL, write.table.args = NULL)
}
\arguments{
\item{infile}{unsorted file to read from. See \code{\link[utils]{read.table}}.}

\item{outfile}{where to write the sorted file. See
\code{\link[utils]{write.table}}. If \code{infile} is the name of a file
then the default prepends "sorted_" to this name.}

\item{sortcolumn}{which column of the data frame to sort on}

\item{nrows}{number of rows in the data.frame held in memory}

\item{read.table.args}{named list of extra arguments to read.table}

\item{write.table.args}{named list of extra arguments to write.table.
Defaults to using read.table.args to preserve the original formatting.}
}
\description{
This function was designed to handle files larger than memory. At most
\code{nrows} will be present in memory at once.
}
