\name{ptMEinit,ptMEinitSrvrs,ptMEinitCons,ptMEsend,ptMErecv,ptMEclose,
ptMEtest,ptMEtestWrkr}
\alias{ptMEinit}
\alias{ptMEinitSrvrs}
\alias{ptMEinitCons}
\alias{ptMEsend}
\alias{ptMErecv}
\alias{ptMEclose}
\alias{ptMEtest}
\alias{ptMEtestWrkr}

\title{
Message-passing utilities.
}

\description{Simple MPI-like functions.}

\usage{
ptMEinit(cls)
ptMEinitSrvrs(cls)
}

\arguments{
  \item{cls}{A cluster for the \pkg{parallel} package.}
}

\details{ }

\examples{
}

\author{
Robin Yancey, Norm Matloff
}

