% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disksort.R
\name{sortbin}
\alias{sortbin}
\title{Sort Bin And Write To Outfile}
\usage{
sortbin(fname, sortcolumn, outfile, nchunks)
}
\arguments{
\item{fname}{name of an intermediate file}

\item{sortcolumn}{See \code{\link{disksort}}}

\item{outfile}{See \code{\link{disksort}}.}

\item{nchunks}{total number of chunks expected}
}
\description{
Last step of disksort
}
