% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disksort.R
\name{writechunk}
\alias{writechunk}
\title{Write Chunk Into Bins}
\usage{
writechunk(chunk, bin_names, bin_files, breaks, sortcolumn)
}
\arguments{
\item{chunk}{\code{data.frame} to be binned}

\item{bin_names}{names for each bin. Useful for debugging}

\item{bin_files}{list of files opened in binary append mode}

\item{breaks}{defines the bins}

\item{sortcolumn}{column determining the bin}
}
\description{
Intermediate step in disksort.
}
