\name{hqs,hqsTest)
\alias{hqs}
\alias{hqsTest)

\title{Distributed Sort}

\description{Sort a distributed vector.}

\usage{
ptMEinit(cls)
ptMEinitSrvrs(cls)
}

\arguments{
  \item{cls}{A cluster for the \pkg{parallel} package.}
}

\details{ 
This system of functions implements a message-passing system, similar to
MPI but much simpler and without the configuration difficulties.
}

\examples{
}

\author{
Robin Yancey, Norm Matloff
}

